feat: stage 2 -- jackpot reward

+++ src/main/java/org/sporty/jackpot/messaging/event/BetCreatedEvent.java
@@ -1,4 +1,6 @@
 package org.sporty.jackpot.messaging.event;
 
-public record BetCreatedEvent(String betId, String userId, String jackpotId, Integer betAmount) {
+import java.math.BigDecimal;
+
+public record BetCreatedEvent(String betId, String userId, String jackpotId, BigDecimal betAmount) {
 }
