+++ src/main/java/org/sporty/jackpot/service/command/CreateBetCommand.java
@@ -1,4 +1,6 @@
 package org.sporty.jackpot.service.command;
 
-public record CreateBetCommand(String betId, String userId, String jackpotId, Integer betAmount) {
+import java.math.BigDecimal;
+
+public record CreateBetCommand(String betId, String userId, String jackpotId, BigDecimal betAmount) {
 }
